+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.3
+- **Set Value** improvements:
+  - Handle explicitly more error types for better UX.
+  - Do not set special `$...` variables.
+
 ## 0.3.2
 - Add `tracciatto.reattachDelay` setting to control delay before reattaching after a debug **Restart** call.
 - Make `tracciatto.patchMaxInspectedValueLength` safer to use.